Rename base class from Cop to Base
This change is just so we follow the new naming pattern of RuboCop, who
recently changd from `RuboCop::Cop::Cop` to `RuboCop::Cop::Base`. Here,
we change `RuboCop::Cop::RSpec::Cop` to `RuboCop::Cop::RSpec::Base`.

To make the commits easier to review, the superclass of all existing
cops will be changed in a separate commit.